requirements.txt
PyQt5
mss
pyaudio
ffmpeg-python
numpy
opencv-python
pyqt5-tools